requirements/requirements.chroma.dev.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
asttokens==3.0.0
backoff==2.2.1
bcrypt==4.2.1
black==24.10.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
coloredlogs==15.0.1
decorator==5.1.1
Deprecated==1.2.15
durationpy==0.9
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.12.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
ipython==8.31.0
jedi==0.19.2
kubernetes==31.0.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
mypy-extensions==1.0.0
numpy==2.2.1
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.12
overrides==7.7.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.7.4
prompt_toolkit==3.0.48
protobuf==5.29.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
setuptools==69.0.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.20.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
wheel==0.42.0
wrapt==1.17.0
zipp==3.21.0
